% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tspecies.R
\name{Tspecies}
\alias{Tspecies}
\title{Correct Tspecies by Ks-distribution's Variance}
\usage{
Tspecies(Ks, miu)
}
\arguments{
\item{Ks}{A vector value representing d/L.}

\item{miu}{An numeric value representing neutral mutation rate.}
}
\value{
A numeric value representing the divergence time of species in generations.
}
\description{
Tspecies uses a pre-fitted GAM model to predict Ne based on
a given Ks value and a neutral mutation rate miu. The Ks refers to a vector
representing d/L, where d is the base difference between sequences, and L
is the sequence length. We will corrected the Ks value automatically by using
multiple substitution and 2Ne based on our data.
}
\examples{
Tspecies(c(0.257, 0.202, 0.066, 0.197, 0.278, 0.089), 0.00000008)

}
